# Conflicts:
#	Client/Default/Client.vcxproj.filters
#	Client/Public/CameraMan_Targeter.h
#	Client/Public/Client_EventDefine.h
#	Engine/Private/Physics_Module.cpp
#	Tool/Private/ImGui_Dockspace_TabBar.cpp
#	Tool/Public/ImGui_Dockspace_TabBar.h

+++ Engine/Private/EffectHandler.cpp
@@ -25,6 +25,7 @@ CEffectHandler::CEffectHandler(const CEffectHandler& rhs)
 
 HRESULT CEffectHandler::Initialize_Prototype(void* pArg)
 {
+
     if(FAILED(Ready_Desc(pArg)))
         return E_FAIL;
 
@@ -50,7 +51,6 @@ HRESULT CEffectHandler::Ready_Desc(void* pArg)
 {
     if (pArg)
         m_tDesc = *(static_cast<ANIM_EFFECT_HANDLER_DESC*>(pArg));
-
     else
         return E_FAIL;
 
@@ -92,7 +92,6 @@ void CEffectHandler::Awake()
 {
     if (m_tDesc.eType == E_HANDLER_TYPE::MODEL_ANIM)
         Ready_AnimState();
-
     else
         Create_SpawnEffect();
 }
@@ -107,7 +106,7 @@ HRESULT CEffectHandler::Gizmo_Setting()
     return S_OK;
 }
 
-void CEffectHandler::GetAnimation()
+void CEffectHandler::Set_OwnerModel()
 {
     if (Get_Owner())
     {
@@ -178,21 +177,12 @@ void CEffectHandler::Ready_Event()
 
 HRESULT CEffectHandler::Ready_AnimState()
 {
-    GetAnimation();
+    Set_OwnerModel();
 
-    if (Get_Owner())
+    if (FAILED(Owner_Setting()))
     {
-        // 툴에서 부모의 Matrix를 참조할 때 안전하게 가져오기
-        auto pTransform = Get_Owner()->Get_Component<CTransform>();
-        if (pTransform)
-        {
-            // PartObject 기반이면 부모 Container의 WorldMatrix를 참조함
-            auto pPart = static_cast<CPartObject*>(Get_Owner());
-            if (pPart->Get_Parent())
-                m_pOwnerMatrix = &pPart->Get_Parent()->Get_Component<CTransform>()->Get_WorldMatrix();
-            else
-                m_pOwnerMatrix = &pTransform->Get_WorldMatrix();
-        }
+        MSG_BOX("Owner Setting Fail : EffectHandler");
+        return E_FAIL;
     }
 
     Ready_Event();
@@ -204,14 +194,12 @@ HRESULT CEffectHandler::Create_SpawnEffect()
 {
     if (m_tDesc.eType == E_HANDLER_TYPE::MODEL_ANIM) return E_FAIL;
 
-    if (m_pOwnerMatrix == nullptr)
-    {
-        auto pPart = static_cast<CPartObject*>(Get_Owner());
-        if (pPart)
-            m_pOwnerMatrix = &pPart->Get_Parent()->Get_Component<CTransform>()->Get_WorldMatrix();
+    m_pOwnerMatrix = m_tDesc.mEffectState[E_OBJ_LIFECYCLE_STATE::ON_SPAWN].pParentTransformMatrix;
 
-        else
-            m_pOwnerMatrix = &pPart->Get_Component<CTransform>()->Get_WorldMatrix();
+    if (FAILED(Owner_Setting()))
+    {
+        MSG_BOX("Owner Setting Fail : EffectHandler");
+        return E_FAIL;
     }
 
     return S_OK;
@@ -222,7 +210,8 @@ void CEffectHandler::Release_Event()
     if (m_tDesc.eType == E_HANDLER_TYPE::MODEL_ANIM)
     {
         if (m_pOwner)
-            m_pOwnerModel->OnNotify.Unsubscribe(m_EventHandle);
+            if(m_pOwnerModel)
+                m_pOwnerModel->OnNotify.Unsubscribe(m_EventHandle);
     }
 }
 
@@ -273,8 +262,25 @@ void CEffectHandler::CallBackEvent(const AnimNotifyKey& key)
     }
 }
 
+HRESULT CEffectHandler::Owner_Setting()
+{
+    // 넣어준 값이 없다면 여기서 설정을 해준다.
+    if (m_pOwnerMatrix == nullptr)
+    {
+        if (dynamic_cast<CPartObject*>(Get_Owner()))
+            m_pOwnerMatrix = &(static_cast<CPartObject*>(Get_Owner())->Get_Parent()->Get_Component<CTransform>()->Get_WorldMatrix());
+
+        else
+            m_pOwnerMatrix = &(Get_Owner()->Get_Component<CTransform>()->Get_WorldMatrix());
+    }
+
+
+    return m_pOwnerMatrix ? S_OK : E_FAIL;
+}
+
 void CEffectHandler::PoolObject_CallBack(CGameObject* pGo)
 {
+    // 콜백 받아온 ActiveEffect를 관리 항목으로 저장한다.
     m_ActiveEffects[ENUM_TO_UINT(m_tDesc.eType)].emplace(pGo->Get_Name(), pGo);
 }
 
@@ -282,33 +288,20 @@ void CEffectHandler::Request_SpawnEffect(const DTO::EFFECTEVENT& Script)
 {
     Matrix MatWorldOffset = XMMatrixIdentity();
     Matrix MatOwnerMatrix = XMMatrixIdentity();
-
     const Matrix* pTargetBoneMatrix = nullptr;
 
-    if (Script.iBoneIndex != -1 && Script.bFollowBone)
+    if (FAILED(Owner_Setting()))
     {
-        if (m_pOwnerModel == nullptr)
-            return;
-
-        pTargetBoneMatrix = &m_pOwnerModel->Get_Bone(Script.iBoneIndex)->Get_CombinedTransformMatrix();
+        MSG_BOX("Owner Setting Fail : EffectHandler");
+        return;
     }
 
-    if (m_pOwnerMatrix == nullptr)
-    {
-        auto pPart = static_cast<CPartObject*>(Get_Owner());
-        if (pPart)
-            m_pOwnerMatrix = &pPart->Get_Parent()->Get_Component<CTransform>()->Get_WorldMatrix();
-
-        else
-            m_pOwnerMatrix = &pPart->Get_Component<CTransform>()->Get_WorldMatrix();
-    }
-
+    // 뼈 행렬 계산
+    BoneMatrix_CalCulator(Script, pTargetBoneMatrix);
     // 애니메이션 모델 떄문에 Scale 행렬을 날려주는 값을 넣어준다.
     MatOwnerMatrix = Delete_ScaleMatrix(*m_pOwnerMatrix);
-
     // 오프셋 적용
     MatWorldOffset = Offset_CalCulator(Script);
-
     // 이펙트 생성 요청
     Spawn_RequestFromEffectManager(Script, MatWorldOffset, MatOwnerMatrix, pTargetBoneMatrix);
 }
@@ -317,30 +310,18 @@ void CEffectHandler::Request_SpawnEffect(const DTO::EFFECTEVENT& script, const s
 {
     Matrix MatWorldOffset = XMMatrixIdentity();
     Matrix MatOwnerMatrix = XMMatrixIdentity();
-
     const Matrix* pTargetBoneMatrix = nullptr;
 
-    if (script.iBoneIndex != -1 && script.bFollowBone)
+    if (FAILED(Owner_Setting()))
     {
-        if (m_pOwnerModel == nullptr)
-            return;
-
-        pTargetBoneMatrix = &m_pOwnerModel->Get_Bone(script.iBoneIndex)->Get_CombinedTransformMatrix();
+        MSG_BOX("Owner Setting Fail : EffectHandler");
+        return;
     }
 
-    if (m_pOwnerMatrix == nullptr)
-    {
-        auto pPart = static_cast<CPartObject*>(Get_Owner());
-        if(pPart)
-            m_pOwnerMatrix = &pPart->Get_Parent()->Get_Component<CTransform>()->Get_WorldMatrix();
-
-        else
-            m_pOwnerMatrix = &pPart->Get_Component<CTransform>()->Get_WorldMatrix();
-    }
-
+    // 뼈 행렬 계산
+    BoneMatrix_CalCulator(script, pTargetBoneMatrix);
     // 애니메이션 모델 떄문에 Scale 행렬을 날려주는 값을 넣어준다.
     MatOwnerMatrix = Delete_ScaleMatrix(*m_pOwnerMatrix);
-
     // 오프셋 적용
     MatWorldOffset = Offset_CalCulator(script);
 
@@ -460,6 +441,17 @@ DTO::EFFECTEVENT CEffectHandler::Write_EffectEventDesc(const E_OBJ_LIFECYCLE_STA
     pEvent.vRotation = pDesc.vRotation;
 
     return pEvent;
+}
+
+void CEffectHandler::BoneMatrix_CalCulator(const DTO::EFFECTEVENT& script, OUT const SimpleMath::Matrix*& BoneMatrix)
+{
+    if (script.iBoneIndex != -1 && script.bFollowBone)
+    {
+        if (m_pOwnerModel == nullptr)
+            return;
+
+        (BoneMatrix) = &m_pOwnerModel->Get_Bone(script.iBoneIndex)->Get_CombinedTransformMatrix();
+    }
 }
 
 SimpleMath::Matrix CEffectHandler::Offset_CalCulator(const DTO::EFFECTEVENT& script)
